Prepare to release version 22: Allow tags to be prefixes with "v"
We're changing to a new versioning scheme, dropping the "0." from the
version numbers as it is redundant and misleading WRT how much we care
about backwards compatibility, etc.  Future tags will take the form
"v22", "v23", etc. rather than "0.22".  The "v" prefix allows the tag
name to be distinguished from a git SHA.